requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
filelock==3.12.2
Flask==2.2.4
Flask==2.2.4
Flask-Cors==4.0.0
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
packaging==23.1
platformdirs==3.8.0
pluggy==1.2.0
psycopg2==2.9.6
PyJWT==2.8.0
pymongo==4.4.1
pytest==7.4.0
python-dotenv==0.21.1
typing_extensions==4.7.1
virtualenv==20.23.1
Werkzeug==1.0.1
